+++ News/app/src/main/java/com/eugenebrusov/news/data/model/NewsItem.kt
@@ -1,4 +1,4 @@
-package com.eugenebrusov.news.data
+package com.eugenebrusov.news.data.model
 
 import android.arch.persistence.room.ColumnInfo
 import android.arch.persistence.room.Entity
