+++ News/app/src/main/java/com/eugenebrusov/news/data/source/DataSource.kt
@@ -1,6 +1,6 @@
 package com.eugenebrusov.news.data.source
 
-import com.eugenebrusov.news.data.NewsItem
+import com.eugenebrusov.news.data.model.NewsItem
 
 /**
  * Main entry point for accessing news data
