+++ News/app/src/main/java/com/eugenebrusov/news/data/source/local/Database.kt
@@ -4,7 +4,7 @@ import android.arch.persistence.room.Database
 import android.arch.persistence.room.Room
 import android.arch.persistence.room.RoomDatabase
 import android.content.Context
-import com.eugenebrusov.news.data.NewsItem
+import com.eugenebrusov.news.data.model.NewsItem
 
 /**
  * The Room Database that contains the NewsItem table.
