+++ News/app/src/main/res/layout/item_news_list.xml
@@ -7,7 +7,7 @@
         <import type="android.view.View.OnClickListener" />
 
         <variable name="newsItem"
-            type="com.eugenebrusov.news.data.NewsItem" />
+            type="com.eugenebrusov.news.data.model.NewsItem" />
 
         <variable name="listener"
             type="com.eugenebrusov.news.newslist.NewsItemUserActionsListener" />
