+++ News/app/src/test/java/com/eugenebrusov/news/data/source/RepositoryTest.kt
@@ -1,6 +1,6 @@
 package com.eugenebrusov.news.data.source
 
-import com.eugenebrusov.news.data.NewsItem
+import com.eugenebrusov.news.data.model.NewsItem
 import com.eugenebrusov.news.newslist.util.any
 import com.eugenebrusov.news.newslist.util.capture
 import org.junit.After
